+++ apps/nextjs-website/src/components/MegaHeader.tsx
@@ -1,12 +1,10 @@
 'use client';
 import React from 'react';
-import {
-  defaultMenuItems,
-  defaultSocialMediaLinks,
-} from '../../stories/MegaHeader/light.stories';
+
 import MegaHeaderRC from '@react-components/components/MegaHeader/MegaHeader';
 import { MegaHeaderProps } from '@react-components/types/MegaHeader/MegaHeader.types';
 import { MegaHeaderSection } from '@/lib/fetch/types/PageSection';
+import { defaultMenuItems, defaultSocialMediaLinks } from '../../stories/MegaHeader/megaheaderCommon';
 
 const makeMegaHeaderProps = ({
   menuItems = defaultMenuItems,
